requirements.txt
numpy>=1.23.4
matplotlib>=3.5.3
pyserial>=3.5
seaborn>=0.12.0